+++ src/engine/Block.h
@@ -8,18 +8,18 @@
 class Block {
 private:
     const static std::vector<Vertex> BLOCK_VERTICES;
-    const static std::vector<uint16_t> BLOCK_INDICES;
+    const static std::vector<uint32_t> BLOCK_INDICES;
 
     std::vector<Vertex> vertices;
-    std::vector<uint16_t> indices;
+    std::vector<uint32_t> indices;
 
     glm::vec3 position{};
 
 public:
-    explicit Block(const glm::vec3 &spawnPosition, uint16_t indexOffset);
+    explicit Block(const glm::vec3 &spawnPosition, uint32_t indexOffset);
 
     [[nodiscard]] std::vector<Vertex> GetVertices() const;
-    [[nodiscard]] std::vector<uint16_t> GetIndices() const;
+    [[nodiscard]] std::vector<uint32_t> GetIndices() const;
 };
 
 #endif //VOXEL_ENGINE_BLOCK_H
